refactor: Extract CSS and composables for better code organization
- Extract KanbanBoard.vue styles to separate KanbanBoard.css file
- Extract SessionListView.vue styles to separate SessionListView.css file
- Move useSessionSubscription from useWebSocket.js to dedicated composable
- Move useProjectSubscription from useWebSocket.js to dedicated composable
- Add new useGlobalSessionSubscription composable for global session updates

This improves code maintainability by separating concerns and reducing
file sizes. The composables can now be imported and tested independently.
