+++ packages/web/src/components/KanbanBoard.vue
@@ -185,6 +185,7 @@ import { useKanbanStore } from '../stores/kanban.js';
 import AddSessionToLaneModal from './AddSessionToLaneModal.vue';
 import LaneSettingsModal from './LaneSettingsModal.vue';
 import MoveCardModal from './MoveCardModal.vue';
+import './KanbanBoard.css';
 
 const props = defineProps({
   projectId: {
@@ -366,404 +367,3 @@ onMounted(() => {
   }
 });
 </script>
-
-<style scoped>
-.kanban-board {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.kanban-loading,
-.kanban-error,
-.kanban-empty {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 2rem;
-  color: var(--color-text-soft);
-  gap: 1rem;
-}
-
-.loading-spinner {
-  width: 24px;
-  height: 24px;
-  border: 2px solid var(--color-border);
-  border-top-color: var(--color-primary);
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.error-icon {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 24px;
-  height: 24px;
-  background-color: var(--color-warning);
-  color: white;
-  border-radius: 50%;
-  font-weight: bold;
-}
-
-.retry-btn {
-  background: none;
-  border: none;
-  color: var(--color-primary);
-  cursor: pointer;
-  padding: 0.5rem 1rem;
-}
-
-.retry-btn:hover {
-  text-decoration: underline;
-}
-
-.kanban-lanes-container {
-  display: flex;
-  gap: 1rem;
-  padding: 1rem;
-  overflow-x: auto;
-  flex: 1;
-  align-items: flex-start;
-}
-
-.kanban-lane {
-  flex-shrink: 0;
-  width: 280px;
-  background: var(--color-bg-soft, rgba(255, 255, 255, 0.05));
-  border-radius: 8px;
-  display: flex;
-  flex-direction: column;
-  max-height: calc(100vh - 200px);
-}
-
-.lane-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0.75rem 1rem;
-  border-bottom: 1px solid var(--color-border);
-}
-
-.lane-title-row {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  flex: 1;
-  min-width: 0;
-}
-
-.lane-title {
-  margin: 0;
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: var(--color-text);
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.lane-automation-indicator {
-  color: #f59e0b;
-  flex-shrink: 0;
-}
-
-.lane-header-actions {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  flex-shrink: 0;
-}
-
-.lane-count {
-  font-size: 0.75rem;
-  color: var(--color-text-soft);
-  background: var(--color-border);
-  padding: 0.125rem 0.5rem;
-  border-radius: 10px;
-}
-
-.lane-settings-btn {
-  background: none;
-  border: none;
-  color: var(--color-text-soft);
-  cursor: pointer;
-  padding: 0.25rem;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  opacity: 0.6;
-  transition: opacity 0.15s, color 0.15s, background-color 0.15s;
-}
-
-.kanban-lane:hover .lane-settings-btn {
-  opacity: 1;
-}
-
-.lane-settings-btn:hover {
-  color: var(--color-primary);
-  background: var(--color-bg-soft);
-}
-
-.lane-cards {
-  flex: 1;
-  overflow-y: auto;
-  padding: 0.5rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  min-height: 100px;
-}
-
-.lane-empty {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 2rem 1rem;
-  color: var(--color-text-soft);
-  font-size: 0.75rem;
-  border: 2px dashed var(--color-border);
-  border-radius: 6px;
-  opacity: 0.6;
-}
-
-.kanban-card {
-  position: relative;
-  background: var(--color-background);
-  border: 1px solid var(--color-border);
-  border-radius: 6px;
-  cursor: grab;
-  transition: border-color 0.2s, box-shadow 0.2s;
-}
-
-.kanban-card:hover {
-  border-color: var(--color-primary);
-}
-
-.kanban-card:active {
-  cursor: grabbing;
-}
-
-.card-link {
-  display: block;
-  padding: 0.75rem;
-  color: var(--color-text);
-  text-decoration: none;
-}
-
-.card-header {
-  display: flex;
-  align-items: flex-start;
-  gap: 0.5rem;
-  margin-bottom: 0.5rem;
-}
-
-.card-status {
-  flex-shrink: 0;
-  font-size: 0.75rem;
-}
-
-.card-status.status-running,
-.card-status.status-starting {
-  color: #10b981;
-}
-
-.card-status.status-waiting {
-  color: #f59e0b;
-}
-
-.card-status.status-completed {
-  color: #3b82f6;
-}
-
-.card-status.status-error,
-.card-status.status-stopped {
-  color: #ef4444;
-}
-
-.card-status.status-scheduled {
-  color: #6b7280;
-}
-
-.card-title {
-  margin: 0;
-  font-size: 0.875rem;
-  font-weight: 500;
-  line-height: 1.3;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-}
-
-.card-meta {
-  display: flex;
-  gap: 0.5rem;
-  font-size: 0.75rem;
-  color: var(--color-text-soft);
-}
-
-.card-mode {
-  text-transform: capitalize;
-}
-
-.card-move-btn {
-  position: absolute;
-  top: 0.25rem;
-  right: 2rem;
-  background: none;
-  border: none;
-  color: var(--color-text-soft);
-  cursor: pointer;
-  opacity: 0;
-  padding: 0.25rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: opacity 0.2s, color 0.2s;
-}
-
-.kanban-card:hover .card-move-btn {
-  opacity: 1;
-}
-
-.card-move-btn:hover {
-  color: var(--color-primary);
-}
-
-.card-remove-btn {
-  position: absolute;
-  top: 0.25rem;
-  right: 0.25rem;
-  background: none;
-  border: none;
-  color: var(--color-text-soft);
-  cursor: pointer;
-  opacity: 0;
-  padding: 0.25rem;
-  font-size: 1rem;
-  line-height: 1;
-  transition: opacity 0.2s, color 0.2s;
-}
-
-.kanban-card:hover .card-remove-btn {
-  opacity: 1;
-}
-
-.card-remove-btn:hover {
-  color: var(--color-danger, #ef4444);
-}
-
-.lane-footer {
-  padding: 0.5rem;
-  border-top: 1px solid var(--color-border);
-  font-size: 0.7rem;
-  color: var(--color-text-soft);
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.lane-automation {
-  display: flex;
-  align-items: center;
-  gap: 0.25rem;
-  padding: 0 0.5rem;
-}
-
-.add-session-btn {
-  width: 100%;
-  padding: 0.5rem;
-  background: transparent;
-  border: 1px dashed var(--color-border);
-  border-radius: 4px;
-  color: var(--color-text-soft);
-  cursor: pointer;
-  font-size: 0.75rem;
-  transition: border-color 0.15s, color 0.15s, background-color 0.15s;
-}
-
-.add-session-btn:hover {
-  border-color: var(--color-primary);
-  color: var(--color-primary);
-  background: rgba(34, 197, 255, 0.05);
-}
-
-.add-lane-container {
-  flex-shrink: 0;
-  min-width: 200px;
-}
-
-.add-lane-btn {
-  width: 100%;
-  padding: 0.75rem 1rem;
-  background: var(--color-bg-soft, rgba(255, 255, 255, 0.05));
-  border: 2px dashed var(--color-border);
-  border-radius: 8px;
-  color: var(--color-text-soft);
-  cursor: pointer;
-  transition: border-color 0.2s, color 0.2s;
-}
-
-.add-lane-btn:hover {
-  border-color: var(--color-primary);
-  color: var(--color-primary);
-}
-
-.add-lane-form {
-  background: var(--color-bg-soft, rgba(255, 255, 255, 0.05));
-  border: 1px solid var(--color-border);
-  border-radius: 8px;
-  padding: 0.75rem;
-}
-
-.add-lane-input {
-  width: 100%;
-  padding: 0.5rem;
-  background: var(--color-background);
-  border: 1px solid var(--color-border);
-  border-radius: 4px;
-  color: var(--color-text);
-  font-size: 0.875rem;
-  margin-bottom: 0.5rem;
-}
-
-.add-lane-input:focus {
-  outline: none;
-  border-color: var(--color-primary);
-}
-
-.add-lane-actions {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.add-lane-confirm,
-.add-lane-cancel {
-  flex: 1;
-  padding: 0.375rem 0.75rem;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 0.75rem;
-}
-
-.add-lane-confirm {
-  background: var(--color-primary);
-  color: white;
-}
-
-.add-lane-cancel {
-  background: var(--color-border);
-  color: var(--color-text);
-}
-</style>
