+++ packages/web/src/views/SessionListView.vue
@@ -237,6 +237,7 @@ import TemplatesPanel from '../components/TemplatesPanel.vue';
 import CommandButtonsPanel from '../components/CommandButtonsPanel.vue';
 import KanbanBoard from '../components/KanbanBoard.vue';
 import AddSessionToLaneModal from '../components/AddSessionToLaneModal.vue';
+import './SessionListView.css';
 
 const route = useRoute();
 const router = useRouter();
@@ -437,327 +438,3 @@ onUnmounted(() => {
   clearTimeout(fetchSummariesTimer);
 });
 </script>
-
-<style scoped>
-.page-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  margin-bottom: 1rem;
-}
-
-.page-header h1 {
-  margin: 0;
-}
-
-.project-title {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.repo-link {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 1.5rem;
-  height: 1.5rem;
-  color: var(--color-primary);
-  transition: color 0.15s, transform 0.15s;
-  padding: 0.25rem;
-}
-
-.repo-link:hover {
-  color: var(--color-primary-bright, #06ffff);
-  transform: scale(1.1);
-}
-
-.repo-icon {
-  width: 100%;
-  height: 100%;
-  stroke-linecap: round;
-  stroke-linejoin: round;
-}
-
-.tabs {
-  display: flex;
-  gap: 0;
-  border-bottom: 1px solid var(--color-border);
-  margin-bottom: 1.5rem;
-  /* Sticky positioning - accounts for header + safe area on iOS */
-  position: -webkit-sticky; /* Safari prefix */
-  position: sticky;
-  /* Default for phones/small screens */
-  top: calc(var(--header-height-computed, 51px) + var(--viewport-offset-top, 0px));
-  background-color: var(--color-background);
-  z-index: 99;
-  padding: 0.5rem 0.5rem 0 0.5rem;
-}
-
-/* iPad and larger screens - account for safe area */
-@media (min-width: 768px) {
-  .tabs {
-    top: calc(var(--header-height-computed, 80px) + var(--viewport-offset-top, 0px));
-  }
-}
-
-.tabs-desktop {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.tabs-mobile {
-  display: none;
-  width: 100%;
-}
-
-.tab-select {
-  width: 100%;
-  padding: 0.75rem;
-  font-size: 0.9rem;
-  background: var(--color-bg);
-  color: var(--color-text);
-  border: 1px solid var(--color-border);
-  border-radius: var(--border-radius);
-  cursor: pointer;
-}
-
-.tabs-left {
-  display: flex;
-  gap: 0;
-}
-
-.tab {
-  background: none;
-  border: none;
-  padding: 0.75rem 1.25rem;
-  font-size: 0.9rem;
-  color: var(--color-text-soft);
-  cursor: pointer;
-  border-bottom: 2px solid transparent;
-  margin-bottom: -1px;
-  transition: color 0.15s, border-color 0.15s;
-}
-
-.tab:hover {
-  color: var(--color-text);
-}
-
-.tab.active {
-  color: var(--color-primary);
-  border-bottom-color: var(--color-primary);
-  font-weight: 500;
-}
-
-.skeleton-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.error-message {
-  color: var(--color-error);
-  padding: 1rem;
-  background-color: rgba(248, 81, 73, 0.1);
-  border-radius: var(--border-radius);
-}
-
-.empty-state {
-  text-align: center;
-  padding: 3rem;
-  color: var(--color-text-soft);
-}
-
-.empty-state p {
-  margin-bottom: 1rem;
-}
-
-.session-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.tab-spacer {
-  height: 1rem;
-}
-
-@media (max-width: 480px) {
-  .page-header {
-    flex-direction: column;
-    gap: 1rem;
-  }
-
-  .page-header .btn {
-    width: 100%;
-    justify-content: center;
-  }
-
-  /* Hide desktop tabs on mobile */
-  .tabs-desktop {
-    display: none !important;
-  }
-
-  /* Show mobile tabs on mobile */
-  .tabs-mobile {
-    display: block;
-  }
-
-  /* Hide desktop-only elements on mobile */
-  .desktop-only {
-    display: none !important;
-  }
-}
-
-/* Responsive utility classes */
-.mobile-only {
-  display: block;
-}
-
-/* Hide desktop-only elements by default (mobile) */
-.desktop-only {
-  display: none;
-}
-
-/* Hide mobile-only elements on desktop */
-@media (min-width: 481px) {
-  .mobile-only {
-    display: none !important;
-  }
-}
-
-/* Show desktop-only elements on desktop */
-@media (min-width: 481px) {
-  .desktop-only {
-    display: inline-flex !important;
-  }
-}
-
-/* Lane Selector Modal */
-.modal-backdrop {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.7);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000;
-}
-
-.modal-content {
-  background: var(--color-background-secondary, #1f2937);
-  border-radius: 0.5rem;
-  width: 100%;
-  max-width: 400px;
-  max-height: 90vh;
-  overflow-y: auto;
-  border: 1px solid var(--color-border);
-}
-
-.modal-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 1.5rem;
-  border-bottom: 1px solid var(--color-border);
-}
-
-.modal-title {
-  margin: 0;
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: var(--color-text);
-}
-
-.close-btn {
-  background: none;
-  border: none;
-  font-size: 1.5rem;
-  color: var(--color-text-soft);
-  cursor: pointer;
-  padding: 0;
-  line-height: 1;
-}
-
-.close-btn:hover {
-  color: var(--color-text);
-}
-
-.modal-body {
-  padding: 1.5rem;
-}
-
-.modal-description {
-  margin: 0 0 1rem;
-  font-size: 0.9rem;
-  color: var(--color-text-soft);
-}
-
-.modal-footer {
-  display: flex;
-  justify-content: flex-end;
-  gap: 0.75rem;
-  padding: 1rem 1.5rem;
-  border-top: 1px solid var(--color-border);
-}
-
-.lane-options {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.lane-option-btn {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem 1rem;
-  background: var(--color-background);
-  border: 1px solid var(--color-border);
-  border-radius: 0.375rem;
-  color: var(--color-text);
-  cursor: pointer;
-  transition: border-color 0.15s, background-color 0.15s;
-}
-
-.lane-option-btn:hover {
-  border-color: var(--color-primary);
-  background: rgba(34, 197, 255, 0.05);
-}
-
-.lane-option-name {
-  font-weight: 500;
-}
-
-.lane-option-count {
-  font-size: 0.75rem;
-  color: var(--color-text-soft);
-}
-
-.empty-lanes {
-  text-align: center;
-  color: var(--color-text-soft);
-  font-size: 0.875rem;
-}
-
-.btn {
-  padding: 0.5rem 1rem;
-  border-radius: 0.25rem;
-  font-weight: 500;
-  cursor: pointer;
-  border: none;
-  font-size: 0.9rem;
-}
-
-.btn-secondary {
-  background: var(--color-background);
-  border: 1px solid var(--color-border);
-  color: var(--color-text);
-}
-
-.btn-secondary:hover {
-  background: var(--color-background-secondary);
-}
-</style>
